attack/requirements.txt
requests==2.27.1
google-cloud==0.34.0
locust==2.8.3
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
ratelimit==2.2.1
urllib3==1.26.8
kubernetes==23.3.0
numpy==1.22.2
gevent==21.12.0
pyfiglet==0.8.post1
python-dateutil~=2.8.2
